drivers: interrupt_controller: add MIWU driver support in NPCM4XX.
The device Multi-Input Wake-Up Unit (MIWU) supports NPCM4XX to exit
'Sleep' or 'Deep Sleep' power state.
Also, it provides signal conditioning and grouping of external
interrupt sources. NPCM4XX has three identical MIWU modules: MIWU0,
MIWU1, MIWU2. Together, they support a total of 143 internal
and/or external wake-up sources.

In this CL, we use device tree files to present the relationship bewteen
MIWU and the other devices in NPCM4XX. It includes:
1. npcm4xx-miwus-int-map.dtsi: it presents relationship between MIWU
group and NVIC interrupt. Basically it's a 1-to-1 mapping but there
are 2 interrupts in one group as an exception.
2. npcm4xx-miwus-wui-map.dtsi: it presents relationship between input of
MIWU and its source device such as gpio, timer, eSPI VWs and so on.

This CL also includes:
1. Add MIWU device tree declarations.
2. MIWU api function declarations and implementation to configure signal
conditions and callback function mechanism. They can be be classified
into two types. One is for GPIO which connects original gpio callback
implemetation and the other is for generic devices such as timer,
eSPI, and so on.

Signed-off-by: Tyrone Ting <kfting@nuvoton.com>