drivers: pinctrl : add npcm pinctrl driver
Add npcm pinctrl driver.

The pinctrl driver sets pin multiplexing with configurations in device
tree, which consist of registers and their bit values. It also sets the
peripheral pull-up/down and the device's power supply type with
configurations in the same way. Other controls, such as tri-state, are
set with raw data.

There are three tables of configurations generated from the device tree:
- npcm_pinmux_table: pin multiplexing configurations.
- npcm_periph_pupd_table: pull-up/down configurations.
- npcm_pin_lv_table: power supply configurations.

Signed-off-by: James Chiang <cpchiang1@nuvoton.com>
Signed-off-by: Alan Yang <tyang1@nuvoton.com>